# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  errors.AppError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  nutrition.CreateNutritionDataRequest:
    properties:
      calcium:
        type: number
      calories:
        type: number
      carbs:
        type: number
      date:
        type: string
      fats:
        type: number
      fiber:
        type: number
      iron:
        type: number
      nutrition_score:
        type: integer
      protein:
        type: number
      sodium:
        type: number
      sugar:
        type: number
      vitamin_a:
        type: number
      vitamin_b:
        type: number
      vitamin_c:
        type: number
      vitamin_d:
        type: number
    type: object
  nutrition.NutritionData:
    properties:
      calcium:
        type: number
      calories:
        type: number
      carbs:
        type: number
      created_at:
        type: string
      date:
        type: string
      fats:
        type: number
      fiber:
        type: number
      id:
        type: string
      iron:
        type: number
      nutrition_score:
        type: integer
      protein:
        type: number
      sodium:
        type: number
      sugar:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
      vitamin_a:
        type: number
      vitamin_b:
        type: number
      vitamin_c:
        type: number
      vitamin_d:
        type: number
    type: object
  nutrition.NutritionStats:
    properties:
      avg_calories:
        type: number
      avg_carbs:
        type: number
      avg_fats:
        type: number
      avg_nutrition_score:
        type: integer
      avg_protein:
        type: number
      total_days:
        type: integer
    type: object
  nutrition.UpdateNutritionDataRequest:
    properties:
      calcium:
        type: number
      calories:
        type: number
      carbs:
        type: number
      fats:
        type: number
      fiber:
        type: number
      iron:
        type: number
      nutrition_score:
        type: integer
      protein:
        type: number
      sodium:
        type: number
      sugar:
        type: number
      vitamin_a:
        type: number
      vitamin_b:
        type: number
      vitamin_c:
        type: number
      vitamin_d:
        type: number
    type: object
...
  preferences.CreatePreferencesRequest:
    properties:
      age_groups:
        additionalProperties: true
        type: object
      allergies:
        items:
          type: string
        type: array
      avoid_excess:
        items:
          type: string
        type: array
      budget_range:
        additionalProperties: true
        type: object
      cooking_frequency:
        type: string
      daily_calories:
        type: integer
      dietary_restrictions:
        items:
          type: string
        type: array
      dietary_type:
        type: string
      eating_schedule:
        additionalProperties: true
        type: object
      health_conditions:
        items:
          type: string
        type: array
      household_id:
        type: string
      household_size:
        minimum: 1
        type: integer
      leftover_comfort_level:
        type: string
      macro_goal:
        additionalProperties: true
        type: object
      meal_prep_preference:
        type: string
      preferred_cuisines:
        items:
          type: string
        type: array
      preferred_stores:
        items:
          type: string
        type: array
      price_sensitivity:
        type: string
      sustainability_preference:
        type: string
      vitamins_focus:
        items:
          type: string
        type: array
      waste_sensitivity_level:
        type: string
      weekly_budget:
        type: number
    required:
    - household_id
    - household_size
    type: object
...
  preferences.FamilyPreferences:
    properties:
      age_groups:
        $ref: '#/definitions/preferences.JSONB'
      allergies:
        items:
          type: string
        type: array
      avoid_excess:
        items:
          type: string
        type: array
      budget_range:
        $ref: '#/definitions/preferences.JSONB'
      cooking_frequency:
        type: string
      created_at:
        type: string
      daily_calories:
        type: integer
      dietary_restrictions:
        items:
          type: string
        type: array
      dietary_type:
        type: string
      eating_schedule:
        $ref: '#/definitions/preferences.JSONB'
      health_conditions:
        items:
          type: string
        type: array
      household_id:
        type: string
      household_size:
        type: integer
      id:
        type: string
      leftover_comfort_level:
        type: string
      macro_goal:
        $ref: '#/definitions/preferences.JSONB'
      meal_prep_preference:
        type: string
      preferred_cuisines:
        items:
          type: string
        type: array
      preferred_stores:
        items:
          type: string
        type: array
      price_sensitivity:
        type: string
      sustainability_preference:
        type: string
      updated_at:
        type: string
      vitamins_focus:
        items:
          type: string
        type: array
      waste_sensitivity_level:
        type: string
      weekly_budget:
        type: number
    type: object
  preferences.JSONB:
    additionalProperties: true
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A Go backend API server for Foodlink application
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Foodlink Backend API
  version: "1.0"
paths:
...
  /nutrition:
    get:
      consumes:
      - application/json
      description: Get nutrition data for the authenticated user within a date range
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/nutrition.NutritionData'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Get nutrition data
      tags:
      - nutrition
    post:
      consumes:
      - application/json
      description: Create a new nutrition data entry
      parameters:
      - description: Nutrition data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/nutrition.CreateNutritionDataRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/nutrition.NutritionData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Log nutrition data
      tags:
      - nutrition
  /nutrition/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a specific nutrition data entry
      parameters:
      - description: Nutrition Data ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/nutrition.NutritionData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Get nutrition data by ID
      tags:
      - nutrition
    put:
      consumes:
      - application/json
      description: Update an existing nutrition data entry
      parameters:
      - description: Nutrition Data ID
        in: path
        name: id
        required: true
        type: string
      - description: Nutrition data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/nutrition.UpdateNutritionDataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/nutrition.NutritionData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Update nutrition data
      tags:
      - nutrition
  /nutrition/stats:
    get:
      consumes:
      - application/json
      description: Get nutrition statistics for the authenticated user
      parameters:
      - description: 'Number of days (default: 30)'
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/nutrition.NutritionStats'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Get nutrition statistics
      tags:
      - nutrition
  /nutrition/today:
    get:
      consumes:
      - application/json
      description: Get today's nutrition data for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/nutrition.NutritionData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Get today's nutrition
      tags:
      - nutrition
  /preferences:
    get:
      consumes:
      - application/json
      description: Get family preferences for the authenticated user's household
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/preferences.FamilyPreferences'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Get family preferences
      tags:
      - preferences
    post:
      consumes:
      - application/json
      description: Create or update family preferences
      parameters:
      - description: Preferences data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/preferences.CreatePreferencesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/preferences.FamilyPreferences'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/preferences.FamilyPreferences'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Create or update preferences
      tags:
      - preferences
    put:
      consumes:
      - application/json
      description: Create or update family preferences
      parameters:
      - description: Preferences data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/preferences.CreatePreferencesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/preferences.FamilyPreferences'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/preferences.FamilyPreferences'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Create or update preferences
      tags:
      - preferences
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"